document-generation-solution-accelerator/infra/scripts/index_scripts/requirements.txt
azure-storage-file-datalake
openai
pypdf
# pyodbc
tiktoken
msal[broker]==1.31.1
azure-identity
azure-ai-textanalytics
azure-search-documents==11.6.0b3
azure-keyvault-secrets
pandas
# pymssql
datetime